Add store model
