+++ src/app/features/sample-api-feat/store/actions/user.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { MockUserData } from '../../types';
+import { MockUserData } from '../model/user.model';
 
 export const saveUserData = createAction(
   '[User API] Save User Data',
