+++ src/app/features/sample-api-feat/store/model/user.model.ts
@@ -1,0 +1,11 @@
+interface Permission {
+  hasReadable: boolean;
+  hasEditable: boolean;
+  hasPublishable: boolean;
+}
+
+export interface MockUserData {
+  id: string;
+  name: string;
+  permission: Permission;
+}
