--- src/app/features/sample-api-feat/types.ts
@@ -1,9 +1,0 @@
-export interface MockUserData {
-  id: string;
-  name: string;
-  permission: {
-    hasReadable: boolean;
-    hasEditable: boolean;
-    hasPublishable: boolean;
-  };
-}
